+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/templatest/compare/v0.2.0...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- add: `getgroup` and `filtergroup` can receive multiple prefixes
 - Adds `templatest.Registered.getbyname`
 
 [0.2.0](https://github.com/jshwi/templatest/releases/tag/v0.2.0) - 2022-04-22